added game menu and disabled network tab in main menu
