+++ src/game/mod.rs
@@ -1,0 +1,22 @@
+mod camera;
+mod input;
+mod local;
+mod music;
+mod network;
+mod scene;
+mod world;
+
+pub use camera::GameCamera;
+pub use local::LocalGame;
+
+pub use scene::{create_game_scene, GameScene};
+
+pub use world::GameWorld;
+
+pub use network::{
+    NetworkConnection, NetworkConnectionKind, NetworkConnectionStatus, NetworkGame, NetworkMessage,
+};
+
+pub use input::{collect_input, GameInput, GameInputScheme};
+
+pub use music::{start_music, stop_music};
