--- src/nodes/mod.rs
@@ -1,13 +1,0 @@
-mod camera;
-mod decoration;
-mod local_network;
-pub mod network;
-mod particle_emitters;
-mod scene_renderer;
-
-pub use camera::Camera;
-pub use decoration::Decoration;
-pub use local_network::LocalNetwork;
-pub use network::Network;
-pub use particle_emitters::ParticleEmitters;
-pub use scene_renderer::SceneRenderer;
